Parser debugging and restructure
