+++ util/defs.hpp
@@ -51,7 +51,6 @@ namespace Util {
 #define member_class(X, Y) \
   private:                 \
   X Y##_;                  \
-  void set_##Y(X x);       \
   public:                  \
   X Y() const;             \
   X &Y##_ref();
